client/requirements.txt
# Client/Backend Requirements

# Core dependencies
fastapi>=0.115.12
uvicorn[standard]>=0.27.1
websockets>=12.0
python-dotenv>=1.0.1
httpx>=0.28.1

# Pydantic AI
pydantic-ai==0.0.15
pydantic-ai-slim[openai,anthropic]==0.0.15

# Database
aiosqlite

# Authentication
email-validator>=2.0.0
python-multipart>=0.0.6

# AI Providers
openai>=1.35.7
anthropic>=0.40.0

# Development dependencies (optional)
# pytest==8.0.1
# pytest-asyncio==0.23.5
# mypy==1.8.0